Fix styling in expenses container
